requirements.txt
pandas==2.2.2
numpy==1.26.4
faker==25.8.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.54
python-dotenv==1.0.1
pyyaml==6.0.1
fastapi==0.111.0
uvicorn==0.30.1
pytest==8.2.2
pytest-cov==5.0.0
apache-airflow==2.9.3
